feat(staking): upgrade amounts to u128

+++ workspace/apps/staking/contracts/src/staking/interface.cairo
@@ -9,20 +9,20 @@ pub struct StakerInfo {
     pub operational_address: ContractAddress,
     pub pooling_contract: Option<ContractAddress>,
     pub unstake_time: Option<felt252>,
-    pub amount_own: u64,
-    pub amount_pool: u64,
-    pub index: u128,
-    pub unclaimed_rewards_own: u64,
-    pub unclaimed_rewards_pool: u64,
+    pub amount_own: u128,
+    pub amount_pool: u128,
+    pub index: u64,
+    pub unclaimed_rewards_own: u128,
+    pub unclaimed_rewards_pool: u128,
 }
 
 
 #[derive(Default, Drop, PartialEq, Serde)]
 pub struct StakingContractInfo {
     pub max_leverage: u64,
-    pub min_stake: u64,
+    pub min_stake: u128,
     pub token_address: ContractAddress,
-    pub global_index: u128,
+    pub global_index: u64,
     pub global_rev_share: u8,
 }
 
@@ -32,26 +32,26 @@ pub trait IStaking<TContractState> {
         ref self: TContractState,
         reward_address: ContractAddress,
         operational_address: ContractAddress,
-        amount: u64,
+        amount: u128,
         pooling_enabled: bool
     ) -> bool;
     fn increase_stake(
-        ref self: TContractState, staker_address: ContractAddress, amount: u64
-    ) -> u64;
-    fn claim_rewards(ref self: TContractState, staker_address: ContractAddress) -> u64;
+        ref self: TContractState, staker_address: ContractAddress, amount: u128
+    ) -> u128;
+    fn claim_rewards(ref self: TContractState, staker_address: ContractAddress) -> u128;
     fn unstake_intent(ref self: TContractState, staker_address: ContractAddress) -> felt252;
-    fn unstake_action(ref self: TContractState, staker_address: ContractAddress) -> u64;
-    fn add_to_pool(ref self: TContractState, staker_address: ContractAddress, amount: u64) -> u64;
+    fn unstake_action(ref self: TContractState, staker_address: ContractAddress) -> u128;
+    fn add_to_pool(ref self: TContractState, staker_address: ContractAddress, amount: u128) -> u128;
     fn remove_from_pool_intent(
-        ref self: TContractState, staker_address: ContractAddress, amount: u64
+        ref self: TContractState, staker_address: ContractAddress, amount: u128
     ) -> felt252;
-    fn remove_from_pool_action(ref self: TContractState, staker_address: ContractAddress) -> u64;
+    fn remove_from_pool_action(ref self: TContractState, staker_address: ContractAddress) -> u128;
     fn switch_pool(
         ref self: TContractState,
         from_staker_address: ContractAddress,
         to_staker_address: ContractAddress,
         pool_address: ContractAddress,
-        amount: u64,
+        amount: u128,
         data: ByteArray
     ) -> bool;
     fn change_reward_address(ref self: TContractState, reward_address: ContractAddress) -> bool;
